refactor(jsi): move the V8 engine layer into the shared tree
Same split as QuickJS: the `nativescript::engine` classes wrapping V8 move to
jsi/v8/ as .cpp, the Apple bridge's Foundation/objc/Mach-O/metadata preamble
stays behind in a header at the original path that includes the neutral one.
No Apple include changed.

V8 is the first engine whose RuntimeState was not already neutral. Its
argument-marshalling caches -- which memoise "this JS value denotes that native
type / method" -- stored Class and SEL directly. They now hold const void*,
with the casts moved into NativeApiV8Marshalling.mm where the rest of the
Objective-C already lives.

That is a generalisation rather than a workaround: the Android runtime needs
exactly the same cache holding jclass and jmethodID, and all four of these are
pointers. Making the slot opaque is what lets one cache serve both bridges.

Verified: iOS TestRunner on V8, 713 specs, 0 failures. The neutral header also
passes -fsyntax-only as plain C++ with no Foundation in reach.

(cherry picked from commit d40017efec63ed0dc8b5384b30663a8e9a81effb)
